+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Desktop - Better message if there is an error when fetching data from QGIS Server
 * Desktop - Fix link to the documentation about the PostgreSQL search `lizmap_search`
 * Desktop - Always check if the layer is published as WFS when opening the dialog instead of at the end
 * Desktop - Add a new button to generate an HTML table from layer fields, like the auto popup